--- core/src/main/kotlin/behavior/GlobalInteractionBehavior.kt
@@ -1,5 +1,0 @@
-package dev.kord.core.behavior
-
-import dev.kord.core.behavior.interaction.InteractionBehavior
-
-public  interface GlobalInteractionBehavior : InteractionBehavior
