+++ core/src/main/kotlin/behavior/interaction/GlobalInteractionBehavior.kt
@@ -1,0 +1,3 @@
+package dev.kord.core.behavior.interaction
+
+public  interface GlobalInteractionBehavior : InteractionBehavior
